requirements.txt
certifi >= 2020.6.20
six >= 1.10
python_dateutil >= 2.5.3
setuptools >= 21.0.0
urllib3 >= 1.15.1
pyyaml
requests~=2.26.0